pyproject.toml
[project]
name = "tadf"
version = "0.1.0"
description = "TADF building-audit report automation (ehitise auditi aruanne)"
requires-python = ">=3.12"
dependencies = [
  "streamlit>=1.40",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "pydantic>=2.7",
  "docxtpl>=0.19",
  "python-docx>=1.1",
  "pyyaml>=6.0",
  "httpx>=0.27",
  "tenacity>=9.0",
  "pillow>=10.4",
  "exifread>=3.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3",
  "ruff>=0.7",
]
llm = [
  "anthropic>=0.40",
]
sign = [
  "pyasice>=1.1",
]
corpus = [
  "pdfplumber>=0.11",
  "pytesseract>=0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tadf"]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.ruff.lint.per-file-ignores]
"app/pages/*.py" = ["E501"]  # long Russian UI strings — chopping hurts readability

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
